sdk-core: PR suggestions
