+++ packages/sdk-core/src/common/Events.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 interface EventCallback {
     callback: (...args: any[]) => unknown;
     once?: boolean;
